pyproject.toml
[project]
name = "agentic-marketers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.39.0",
    "pillow>=11.3.0",
    "strands-agents>=1.9.1",
    "strands-agents-builder>=0.1.10",
    "strands-agents-tools>=0.2.8",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
]
